Move job code to new job class for jobs. Job well done.
